refactor(memory-server): rename src/ to bot_memory_server/ package
RHCLOUD-48377
Proper Python package structure so relative imports work.
- git mv src/ → bot_memory_server/
- Add [build-system] + setuptools config to pyproject.toml
- Update all test imports from src.* to bot_memory_server.*
- Remove sys.path hacks from test files
- Update Dockerfile, Makefile, docker-compose, vite config
- Update .gitignore paths for dashboard build output
