+++ dashboard/vite.config.ts
@@ -4,7 +4,7 @@ import react from '@vitejs/plugin-react'
 export default defineConfig({
   plugins: [react()],
   build: {
-    outDir: '../memory-server/src/static',
+    outDir: '../memory-server/bot_memory_server/static',
     emptyOutDir: false,
   },
   server: {
