+++ memory-server/bot_memory_server/schema.sql
