+++ memory-server/docker-compose.yml
@@ -23,7 +23,7 @@ services:
     environment:
       DATABASE_URL: postgresql://bot:bot@postgres:5432/bot_memory
     volumes:
-      - ./src/static:/app/src/static
+      - ./bot_memory_server/static:/app/bot_memory_server/static
     depends_on:
       postgres:
         condition: service_healthy
