+++ memory-server/pyproject.toml
@@ -1,3 +1,7 @@
+[build-system]
+requires = ["setuptools>=68.0"]
+build-backend = "setuptools.build_meta"
+
 [project]
 name = "bot-memory-server"
 version = "0.1.0"
@@ -20,3 +24,12 @@ test = [
     "pytest-asyncio>=0.24",
     "httpx>=0.27",
 ]
+
+[tool.setuptools.packages.find]
+include = ["bot_memory_server*"]
+
+[tool.setuptools.package-data]
+bot_memory_server = ["schema.sql", "static/**/*"]
+
+[tool.pytest.ini_options]
+asyncio_mode = "strict"
