+++ memory-server/tests/test_dual_write.py
@@ -6,8 +6,6 @@ alongside the original jira_key/pr_number/pr_url columns.
 
 import json
 import os
-import sys
-from pathlib import Path
 
 import pytest
 
@@ -15,8 +13,7 @@ from conftest import SCHEMA_PATH
 
 os.environ.setdefault("JIRA_URL", "https://redhat.atlassian.net")
 
-sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
-from artifacts import JIRA_BASE_URL, build_artifacts  # noqa: E402
+from bot_memory_server.artifacts import JIRA_BASE_URL, build_artifacts  # noqa: E402
 
 
 async def _apply_schema(db):
